Linux_Agent/requirements-clean.txt
appopener
ddgs
edge_tts
edge_tts
googlesearch_python
groq
keyboard
mtranslate
Pillow
protobuf
psutil
psutil
pygame
PyQt5
PyQt5_sip
pyqt5_sip
python-dotenv
pywhatkit
Requests
rich
selenium
webdriver_manager
